cmd/ore: Move GCloud commands under a subcommand
Don't make them sound more generic than they are, and avoid leaking
GCloud-specific options into unrelated commands.
