+++ cmd/ore/gcloud/image.go
@@ -12,44 +12,43 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package gcloud
 
 import (
 	"fmt"
 	"os"
 
 	"github.com/spf13/cobra"
-
-	"github.com/coreos/mantle/platform/api/gcloud"
 )
 
 var (
-	cmdList = &cobra.Command{
-		Use:   "list-instances -prefix=<prefix>",
-		Short: "List instances on GCE",
-		Run:   runList,
+	cmdImage = &cobra.Command{
+		Use:   "list-images -prefix=<prefix>",
+		Short: "List images in GCE",
+		Run:   runImage,
 	}
+
+	imagePrefix string
 )
 
 func init() {
-	root.AddCommand(cmdList)
+	cmdImage.Flags().StringVar(&imagePrefix, "prefix", "", "prefix to filter list by")
+	GCloud.AddCommand(cmdImage)
 }
 
-func runList(cmd *cobra.Command, args []string) {
+func runImage(cmd *cobra.Command, args []string) {
 	if len(args) != 0 {
 		fmt.Fprintf(os.Stderr, "Unrecognized args in plume list cmd: %v\n", args)
 		os.Exit(2)
 	}
 
-	vms, err := api.ListInstances(opts.BaseName)
+	images, err := api.ListImages(imagePrefix)
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "Failed listing vms: %v\n", err)
+		fmt.Fprintf(os.Stderr, "Failed listing images: %v\n", err)
 		os.Exit(1)
 	}
 
-	for _, vm := range vms {
-		_, extIP := gcloud.InstanceIPs(vm)
-		fmt.Printf("%v:\n", vm.Name)
-		fmt.Printf(" extIP: %v\n", extIP)
+	for _, image := range images {
+		fmt.Printf("%v\n", image)
 	}
 }
